Move kernel ranking to preferred controller component
