+++ src/kernels/kernelFinder.ts
@@ -1,20 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type * as nbformat from '@jupyterlab/nbformat';
-import { inject, injectable } from 'inversify';
+import { injectable } from 'inversify';
 import { CancellationToken } from 'vscode';
 import { Telemetry } from '../platform/common/constants';
 import { Resource } from '../platform/common/types';
 import { StopWatch } from '../platform/common/utils/stopWatch';
-import { traceError, traceDecoratorVerbose, logValue, ignoreLogging } from '../platform/logging';
-import { TraceOptions } from '../platform/logging/types';
-import { PythonEnvironment } from '../platform/pythonEnvironments/info';
-import { captureTelemetry, sendTelemetryEvent } from '../telemetry';
-import { rankKernels, isExactMatch } from './helpers';
+import { sendTelemetryEvent } from '../telemetry';
 import { IContributedKernelFinder } from './internalTypes';
-import { PreferredRemoteKernelIdProvider } from './jupyter/preferredRemoteKernelIdProvider';
-import { IKernelFinder, isLocalConnection, KernelConnectionMetadata } from './types';
+import { IKernelFinder, KernelConnectionMetadata } from './types';
 
 /**
  * Generic class for finding kernels (both remote and local). Handles all of the caching of the results.
@@ -25,48 +19,10 @@ export class KernelFinder implements IKernelFinder {
     private fetchingTelemetrySent = new Set<string>();
     private _finders: IContributedKernelFinder[] = [];
 
-    constructor(
-        @inject(PreferredRemoteKernelIdProvider) private readonly preferredRemoteFinder: PreferredRemoteKernelIdProvider
-    ) {}
+    constructor() {}
 
     public registerKernelFinder(finder: IContributedKernelFinder) {
         this._finders.push(finder);
-    }
-
-    @traceDecoratorVerbose('Rank Kernels', TraceOptions.BeforeCall | TraceOptions.Arguments)
-    @captureTelemetry(Telemetry.RankKernelsPerf)
-    public async rankKernels(
-        resource: Resource,
-        notebookMetadata?: nbformat.INotebookMetadata,
-        @logValue<PythonEnvironment>('uri') preferredInterpreter?: PythonEnvironment,
-        @ignoreLogging() cancelToken?: CancellationToken,
-        useCache?: 'useCache' | 'ignoreCache',
-        serverId?: string
-    ): Promise<KernelConnectionMetadata[] | undefined> {
-        try {
-            // Get list of all of the specs from the cache and without the cache (note, cached items will be validated before being returned)
-            let kernels = await this.listKernels(resource, cancelToken, useCache);
-            if (serverId) {
-                kernels = kernels.filter((kernel) => !isLocalConnection(kernel) && kernel.serverId === serverId);
-            }
-            const preferredRemoteKernelId =
-                resource &&
-                this.preferredRemoteFinder &&
-                this.preferredRemoteFinder.getPreferredRemoteKernelId(resource);
-
-            let rankedKernels = rankKernels(
-                kernels,
-                resource,
-                notebookMetadata,
-                preferredInterpreter,
-                preferredRemoteKernelId
-            );
-
-            return rankedKernels;
-        } catch (ex) {
-            traceError(`RankKernels crashed`, ex);
-            return undefined;
-        }
     }
 
     public async listKernels(
@@ -92,17 +48,6 @@ export class KernelFinder implements IKernelFinder {
         });
 
         return kernels;
-    }
-
-    public isExactMatch(
-        resource: Resource,
-        kernelConnection: KernelConnectionMetadata,
-        notebookMetadata: nbformat.INotebookMetadata | undefined
-    ): boolean {
-        const preferredRemoteKernelId =
-            resource && this.preferredRemoteFinder && this.preferredRemoteFinder.getPreferredRemoteKernelId(resource);
-
-        return isExactMatch(kernelConnection, notebookMetadata, preferredRemoteKernelId);
     }
 
     private finishListingKernels(
